+++ stor/cli.py
@@ -91,8 +91,9 @@ from stor import exceptions
 from stor import settings
 from stor import Path
 from stor import utils
+from stor.extensions import swiftstack
 
-PRINT_CMDS = ('list', 'listdir', 'ls', 'cat', 'pwd', 'walkfiles')
+PRINT_CMDS = ('list', 'listdir', 'ls', 'cat', 'pwd', 'walkfiles', 'url', 'convert-swiftstack')
 SERVICES = ('s3', 'swift')
 
 ENV_FILE = os.path.expanduser('~/.stor-cli.env')
@@ -268,6 +269,25 @@ def get_path(pth, mode=None):
     return prefix / path_part.split(rel_part, depth)[depth].lstrip('/')
 
 
+def _to_url(path):
+    if stor.is_filesystem_path(path):
+        raise ValueError('must be swift or s3 path')
+    return stor.Path(path).to_url()
+
+
+def _convert_swiftstack(path, bucket=None):
+    path = stor.Path(path)
+    if utils.is_swift_path(path):
+        if not bucket:
+            # TODO (jtratner): print help here
+            raise ValueError('--bucket is required for swift paths')
+        return swiftstack.swift_to_s3(path, bucket=bucket)
+    elif utils.is_s3_path(path):
+        return swiftstack.s3_to_swift(path)
+    else:
+        raise ValueError("invalid path for conversion: '%s'" % path)
+
+
 def create_parser():
     parser = argparse.ArgumentParser(description='A command line interface for stor.')
 
@@ -307,7 +327,6 @@ def create_parser():
     parser_ls.set_defaults(func=stor.listdir)
 
     cp_msg = 'Copy a source to a destination path.'
-    cp_msg = 'Alias for copy.'
     parser_cp = subparsers.add_parser('cp',  # noqa
                                       help=cp_msg,
                                       description='%s\n \'-\' is a special character that allows'
@@ -370,6 +389,14 @@ def create_parser():
                                          ' will be cleared if SERVICE is omitted.' % clear_msg)
     parser_clear.add_argument('service', nargs='?', type=str, metavar='SERVICE')
     parser_clear.set_defaults(func=_clear_env)
+    url_parser = subparsers.add_parser('url', help='generate URI for path')
+    url_parser.add_argument('path')
+    url_parser.set_defaults(func=_to_url)
+    parser_swiftstack = subparsers.add_parser('convert-swiftstack',
+                                              help='convert swiftstack paths')
+    parser_swiftstack.add_argument('path')
+    parser_swiftstack.add_argument('--bucket', default=None)
+    parser_swiftstack.set_defaults(func=_convert_swiftstack)
 
     return parser